add sbt-spiewak-sonatype

+++ modules/memcached/src/main/scala/scalacache/memcached/MemcachedTTLConverter.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 scalacache
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package scalacache.memcached
 
 import java.time.{Clock, Instant}
@@ -40,7 +56,7 @@ trait MemcachedTTLConverter {
     case d if d <= 30.days => d.toSeconds.toInt
 
     case d => {
-      val expiryTime = Instant.now(clock).plusSeconds(d.toSeconds.toInt)
+      val expiryTime = Instant.now(clock).plusSeconds(d.toSeconds.toLong)
       (expiryTime.toEpochMilli / 1000).toInt
     }
   }
